+++ src/components/atoms/Link/typings.d.ts
@@ -1,7 +1,10 @@
-interface LinkAttributes extends BlockAttributes {
-	href: string;
-	prefetch: boolean;
-	scroll: boolean;
-}
+interface LinkAttributes extends BlockAttributes {}
 
-interface LinkProps extends LinkAttributes, BlockProps {}
+interface LinkProps
+	extends LinkAttributes,
+		BlockProps,
+		React.HTMLProps<HTMLAnchorElement> {
+	prefetch?: boolean;
+	ref?: Ref<HTMLAnchorElement>;
+	scroll?: boolean;
+}
